+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Data parts can now be emitted based on tool output
 
+### Changed
+
+- Instead of failing when the agent reaches the maximum execution time, a HITL message is thrown asking the user whether to continue
+
 ### Fixed
 
 - Multiple tools with HITL needed in the same round no longer cause an error